#1627: swap tx details
