+++ wallet/cli/cli.cpp
@@ -25,6 +25,7 @@
 #ifdef BEAM_ATOMIC_SWAP_SUPPORT
 #include "wallet/transactions/swaps/common.h"
 #include "wallet/transactions/swaps/utils.h"
+#include "wallet/transactions/swaps/swap_tx_description.h"
 #include "swaps.h"
 #endif // BEAM_ATOMIC_SWAP_SUPPORT
 
@@ -108,6 +109,113 @@ namespace beam
 
 namespace
 {
+    std::string TxDetailsInfo(const IWalletDB::Ptr& walletDB, const TxID& txID)
+    {
+        storage::PaymentInfo pi;
+        auto tx = walletDB->getTx(txID);
+
+        TxDescription desc(*tx);
+        TxAddressType addressType = GetAddressType(desc);
+
+        bool bSuccess = true;
+        bool hasNoPeerId = tx->m_sender && (addressType == TxAddressType::PublicOffline || addressType == TxAddressType::MaxPrivacy);
+        if (!hasNoPeerId)
+        {
+            bSuccess = bSuccess && storage::getTxParameter(*walletDB,
+                txID,
+                tx->m_sender
+                ? TxParameterID::PeerID
+                : TxParameterID::MyID,
+                pi.m_Receiver);
+        }
+
+        bSuccess = bSuccess && storage::getTxParameter(*walletDB,
+            txID,
+            tx->m_sender
+            ? TxParameterID::MyID
+            : TxParameterID::PeerID,
+            pi.m_Sender);
+
+        bSuccess = bSuccess && storage::getTxParameter(*walletDB, txID, TxParameterID::Amount, pi.m_Amount);
+
+        if (addressType == TxAddressType::AtomicSwap)
+        {
+#ifdef BEAM_ATOMIC_SWAP_SUPPORT
+            Amount swapAmount = 0u;
+            AtomicSwapCoin swapCoin = AtomicSwapCoin::Bitcoin;
+            bool isBeamSide = true;
+            bSuccess = bSuccess && storage::getTxParameter(*walletDB, txID, TxParameterID::AtomicSwapAmount, swapAmount);
+            bSuccess = bSuccess && storage::getTxParameter(*walletDB, txID, TxParameterID::AtomicSwapCoin, swapCoin);
+            bSuccess = bSuccess && storage::getTxParameter(*walletDB, txID, TxParameterID::AtomicSwapIsBeamSide, isBeamSide);
+
+            if (bSuccess)
+            {
+                // TODO roman.strilets need to refactor this code
+                auto txDescription = walletDB->getTx(txID);
+                if (txDescription)
+                {
+                    SwapTxDescription swapDescription(*txDescription);
+                    auto swapToken = swapDescription.getToken();
+
+                    if (swapToken)
+                    {
+                        std::ostringstream s;
+
+                        s << "Type:        " << "atomic swap" << std::endl;
+                        s << "Sender:      " << std::to_string(pi.m_Sender) << std::endl;
+                        s << "Swap coin:   " << std::to_string(swapCoin) << std::endl;
+                        s << "Beam side:   " << isBeamSide << std::endl;
+                        s << "Receiver:    " << (hasNoPeerId ? desc.getToken() : std::to_string(pi.m_Receiver)) << std::endl;
+                        s << "Beam amount: " << PrintableAmount(pi.m_Amount) << std::endl;
+                        s << "Swap amount: " << std::to_string(swapAmount) << std::endl;
+                        s << "Swap token:  " << *swapToken << std::endl;
+
+                        return s.str();
+                    }
+                }
+            }
+#endif // BEAM_ATOMIC_SWAP_SUPPORT
+        }
+        else
+        {
+            bSuccess = bSuccess && storage::getTxParameter(*walletDB, txID, TxParameterID::KernelID, pi.m_KernelID);
+
+            if (bSuccess)
+            {
+                auto senderIdentity = tx->getSenderIdentity();
+                auto receiverIdentity = tx->getReceiverIdentity();
+                bool showIdentity = !senderIdentity.empty() && !receiverIdentity.empty();
+
+                auto sender = std::to_string(pi.m_Sender);
+                if (tx->m_txType == wallet::TxType::PushTransaction && !tx->m_sender)
+                {
+                    sender = "shielded pool";
+                }
+
+                std::ostringstream s;
+                s << "Sender: " << sender << std::endl;
+                if (showIdentity)
+                {
+                    s << "Sender identity: " << senderIdentity << std::endl;
+                }
+
+                s << "Receiver: " << (hasNoPeerId ? desc.getToken() : std::to_string(pi.m_Receiver)) << std::endl;
+                if (showIdentity)
+                {
+                    s << "Receiver identity: " << receiverIdentity << std::endl;
+                }
+
+                s << "Amount: " << PrintableAmount(pi.m_Amount) << std::endl;
+                s << "KernelID: " << std::to_string(pi.m_KernelID) << std::endl;
+
+                return s.str();
+            }
+        }
+
+        LOG_WARNING() << "Can't get transaction details";
+        return "";
+    }
+
     SecString GetPassword(const po::variables_map& vm);
 
     void ResolveWID(PeerID& res, const std::string& s)
@@ -1491,7 +1599,7 @@ namespace
             return -1;
         }
 
-        const auto txdetails = storage::TxDetailsInfo(walletDB, *txId);
+        const auto txdetails = TxDetailsInfo(walletDB, *txId);
         if (txdetails.empty()) {
             // storage::TxDetailsInfo already printed an error
             return -1;
