+++ wallet/core/wallet_db.cpp
@@ -5625,71 +5625,6 @@ namespace beam::wallet
             m_KernelID = kernel.m_Internal.m_ID;
         }
 
-        std::string TxDetailsInfo(const IWalletDB::Ptr& walletDB, const TxID& txID)
-        {
-            PaymentInfo pi;
-            auto tx = walletDB->getTx(txID);
-
-            TxDescription desc(*tx);
-            TxAddressType addressType = GetAddressType(desc);
-
-            bool bSuccess = true;
-            bool hasNoPeerId = tx->m_sender && (addressType == TxAddressType::PublicOffline || addressType == TxAddressType::MaxPrivacy);
-            if (!hasNoPeerId)
-            {
-                bSuccess = bSuccess && storage::getTxParameter(*walletDB,
-                                    txID,
-                                    tx->m_sender
-                                        ? TxParameterID::PeerID
-                                        : TxParameterID::MyID,
-                                    pi.m_Receiver);
-            }
-
-            bSuccess = bSuccess && storage::getTxParameter(*walletDB,
-                                txID,
-                                tx->m_sender
-                                    ? TxParameterID::MyID
-                                    : TxParameterID::PeerID,
-                                pi.m_Sender);
-
-            bSuccess = bSuccess && storage::getTxParameter(*walletDB, txID, TxParameterID::KernelID, pi.m_KernelID);
-            bSuccess = bSuccess && storage::getTxParameter(*walletDB, txID, TxParameterID::Amount, pi.m_Amount);
-
-            if (bSuccess)
-            {
-                auto senderIdentity   = tx->getSenderIdentity();
-                auto receiverIdentity = tx->getReceiverIdentity();
-                bool showIdentity     = !senderIdentity.empty() && !receiverIdentity.empty();
-
-                auto sender = std::to_string(pi.m_Sender);
-                if (tx->m_txType == wallet::TxType::PushTransaction && !tx->m_sender)
-                {
-                    sender = "shielded pool";
-                }
-
-                std::ostringstream s;
-                s << "Sender: " << sender << std::endl;
-                if (showIdentity)
-                {
-                    s << "Sender identity: " << senderIdentity << std::endl;
-                }
-
-                s << "Receiver: " << (hasNoPeerId ? desc.getToken() : std::to_string(pi.m_Receiver)) << std::endl;
-                if (showIdentity)
-                {
-                    s << "Receiver identity: " << receiverIdentity << std::endl;
-                }
-
-                s << "Amount: " << PrintableAmount(pi.m_Amount) << std::endl;
-                s << "KernelID: " << std::to_string(pi.m_KernelID) << std::endl;
-
-                return s.str();
-            }
-
-            LOG_WARNING() << "Can't get transaction details";
-            return "";
-        }
-
         namespace
         {
             ByteBuffer ExportSimplePaymentProof(const IWalletDB& walletDB, const TxID& txID)
